Create display
  This file deals with OpenGL functions. In this file I have tried to define window sizes for different files in which I have loaded my virtual object into.I have also assigned textures into the object. The'auxsolidteapot' function displays my teapot and I have also played what is displayed when the camera is opened in real time.

So basically what is happening in this project is that I have created a virtual object and estimated it on my hand by doing its pose estimation. Skin detection is being done and after that pose is estimated and object is moved with respect to my hands movement. Adding on further with my code, I also made a hypothetical screen on my hand with my middle finger , a centroid(basically the centre of the hand) and the thumb. The rectangle on my hand also displays a image that I have approximated inside it.        